Add Version Hook
